Organize precompile session client
