+++ src/tempo/precompile/client/index.ts
@@ -1,2 +1,1 @@
 export * as ChannelOps from './ChannelOps.js'
-export { session } from './Session.js'
