+++ src/tempo/precompile/index.ts
@@ -1,5 +1,6 @@
 export * as Client from './client/index.js'
-export { session } from './client/Session.js'
+export { session } from './session/Client.js'
+export * as Session from './session/index.js'
 export * as Chain from './Chain.js'
 export * as Channel from './Channel.js'
 export * as Constants from './Constants.js'
