+++ src/tempo/precompile/session/Client.test.ts
@@ -8,7 +8,7 @@ import * as Channel from '../Channel.js'
 import { tip20ChannelEscrow } from '../Constants.js'
 import * as Types from '../Types.js'
 import * as Voucher from '../Voucher.js'
-import { session } from './Session.js'
+import { session } from './Client.js'
 
 const account = privateKeyToAccount(
   '0xac0974bec39a17e36ba6a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80',
