+++ src/tempo/precompile/session/index.ts
@@ -1,0 +1,1 @@
+export { session } from './Client.js'
